refact: move z3solver_impl to its own file
