+++ src/main.rs
@@ -14,6 +14,7 @@ use unicorn::unicorn::memory::replace_memory;
 use unicorn::unicorn::optimize::{optimize_model, optimize_model_with_input};
 use unicorn::unicorn::qubot::{InputEvaluator, Qubot};
 use unicorn::unicorn::solver::*;
+use unicorn::unicorn::z3solver_impl;
 use unicorn::unicorn::unroller::{prune_model, renumber_model, unroll_model};
 use unicorn::unicorn::write_model;
 
@@ -81,6 +82,8 @@ fn main() -> Result<()> {
             let emulate_model = is_beator && args.contains_id("emulate");
             let arg0 = expect_arg::<String>(args, "input-file")?;
             let extras = collect_arg_values(args, "extras");
+            println!("arg0: {:?}", arg0);
+
 
             let model = if !input_is_dimacs {
                 let mut model = if !input_is_btor2 {
